Add some global statistics
I know that the desync code is getting triggered some, so this lets us
keep track of how badly it's affecting tracking. In my testing: not bad.
Only 1-2 desyncs per minute, which seems acceptable for now.

For the record: a desync means the connection has gotten into some state
that we don't understand and we have to re-synchronize.